add requested changes
