+++ src/tests/mocks/ubuntuArchiveInfo.ts
@@ -2309,234 +2309,6 @@ export const ubuntuArchiveInfo = {
         },
       ],
     },
-    {
-      slug: "resolute-backports",
-      label: "resolute-backports",
-      preselected: false,
-      components: [
-        {
-          slug: "main",
-          preselected: true,
-        },
-        {
-          slug: "restricted",
-          preselected: false,
-        },
-        {
-          slug: "universe",
-          preselected: false,
-        },
-        {
-          slug: "multiverse",
-          preselected: false,
-        },
-      ],
-      architectures: [
-        {
-          slug: "amd64",
-          preselected: true,
-        },
-        {
-          slug: "amd64v3",
-          preselected: false,
-        },
-        {
-          slug: "arm64",
-          preselected: false,
-        },
-        {
-          slug: "armhf",
-          preselected: false,
-        },
-        {
-          slug: "i386",
-          preselected: false,
-        },
-        {
-          slug: "ppc64el",
-          preselected: false,
-        },
-        {
-          slug: "riscv64",
-          preselected: false,
-        },
-        {
-          slug: "s390x",
-          preselected: false,
-        },
-      ],
-    },
-    {
-      slug: "resolute-proposed",
-      label: "resolute-proposed",
-      preselected: false,
-      components: [
-        {
-          slug: "main",
-          preselected: true,
-        },
-        {
-          slug: "restricted",
-          preselected: false,
-        },
-        {
-          slug: "universe",
-          preselected: false,
-        },
-        {
-          slug: "multiverse",
-          preselected: false,
-        },
-      ],
-      architectures: [
-        {
-          slug: "amd64",
-          preselected: true,
-        },
-        {
-          slug: "amd64v3",
-          preselected: false,
-        },
-        {
-          slug: "arm64",
-          preselected: false,
-        },
-        {
-          slug: "armhf",
-          preselected: false,
-        },
-        {
-          slug: "i386",
-          preselected: false,
-        },
-        {
-          slug: "ppc64el",
-          preselected: false,
-        },
-        {
-          slug: "riscv64",
-          preselected: false,
-        },
-        {
-          slug: "s390x",
-          preselected: false,
-        },
-      ],
-    },
-    {
-      slug: "resolute-security",
-      label: "resolute-security",
-      preselected: true,
-      components: [
-        {
-          slug: "main",
-          preselected: true,
-        },
-        {
-          slug: "restricted",
-          preselected: false,
-        },
-        {
-          slug: "universe",
-          preselected: false,
-        },
-        {
-          slug: "multiverse",
-          preselected: false,
-        },
-      ],
-      architectures: [
-        {
-          slug: "amd64",
-          preselected: true,
-        },
-        {
-          slug: "amd64v3",
-          preselected: false,
-        },
-        {
-          slug: "arm64",
-          preselected: false,
-        },
-        {
-          slug: "armhf",
-          preselected: false,
-        },
-        {
-          slug: "i386",
-          preselected: false,
-        },
-        {
-          slug: "ppc64el",
-          preselected: false,
-        },
-        {
-          slug: "riscv64",
-          preselected: false,
-        },
-        {
-          slug: "s390x",
-          preselected: false,
-        },
-      ],
-    },
-    {
-      slug: "resolute-updates",
-      label: "resolute-updates",
-      preselected: true,
-      components: [
-        {
-          slug: "main",
-          preselected: true,
-        },
-        {
-          slug: "restricted",
-          preselected: false,
-        },
-        {
-          slug: "universe",
-          preselected: false,
-        },
-        {
-          slug: "multiverse",
-          preselected: false,
-        },
-      ],
-      architectures: [
-        {
-          slug: "amd64",
-          preselected: true,
-        },
-        {
-          slug: "amd64v3",
-          preselected: false,
-        },
-        {
-          slug: "arm64",
-          preselected: false,
-        },
-        {
-          slug: "armhf",
-          preselected: false,
-        },
-        {
-          slug: "i386",
-          preselected: false,
-        },
-        {
-          slug: "ppc64el",
-          preselected: false,
-        },
-        {
-          slug: "riscv64",
-          preselected: false,
-        },
-        {
-          slug: "s390x",
-          preselected: false,
-        },
-      ],
-    },
   ],
 } as const satisfies UbuntuArchiveInfo;
