+++ src/tests/server/handlers/publications.ts
@@ -2,28 +2,22 @@ import { API_URL_DEB_ARCHIVE } from "@/constants";
 import type {
   PublishPublicationResponse,
   BatchGetPublicationTargetsResponse,
-  BatchGetMirrorsResponse,
   BatchGetLocalsResponse,
 } from "@/features/publications";
 import { getEndpointStatus } from "@/tests/controllers/controller";
 import {
   locals,
-  mirrors,
   publications,
   publicationTargets,
 } from "@/tests/mocks/publications";
 import type { StrictResponse } from "msw";
 import { delay, http, HttpResponse } from "msw";
-import { generateFilteredResponse } from "./_helpers";
+import {
+  generateFilteredResponse,
+  getDebArchivePaginatedResponse,
+  getDebArchivePaginationParams,
+} from "./_helpers";
 import { ENDPOINT_STATUS_API_ERROR } from "./_constants";
-import type {
-  DeleteMirrorResponse,
-  SyncMirrorResponse,
-  ListMirrorPackagesResponse,
-  GetMirrorResponse,
-  UpdateMirrorResponse,
-  CreateMirrorResponse,
-} from "@canonical/landscape-openapi";
 
 const matchesPublicationsPath = (endpointPath?: string) =>
   !endpointPath || endpointPath.includes("publications");
@@ -46,52 +40,9 @@ const toObjectBody = (value: unknown): Record<string, unknown> => {
   return {};
 };
 
-const getPaginationParams = (requestUrl: string) => {
-  const { searchParams } = new URL(requestUrl);
-  const pageSize = parseInt(searchParams.get("pageSize") ?? "20", 10);
-  const pageToken = parseInt(searchParams.get("pageToken") ?? "0", 10) || 0;
-
-  return {
-    pageSize,
-    pageToken,
-  };
-};
-
-const getPaginatedResponse = <T extends Record<string, unknown>>(
-  data: T[],
-  pageToken: number,
-  pageSize: number,
-) => {
-  const paginatedData = data.slice(pageToken, pageToken + pageSize);
-
-  const nextPageToken =
-    pageToken + pageSize < data.length
-      ? String(pageToken + pageSize)
-      : undefined;
-
-  return {
-    paginatedData,
-    nextPageToken,
-  };
-};
-
-const getMirrorsResponse = (requestUrl: string) => {
-  const { pageSize, pageToken } = getPaginationParams(requestUrl);
-  const { paginatedData, nextPageToken } = getPaginatedResponse(
-    mirrors,
-    pageToken,
-    pageSize,
-  );
-
-  return HttpResponse.json({
-    mirrors: paginatedData,
-    nextPageToken,
-  });
-};
-
 const getLocalsResponse = (requestUrl: string) => {
-  const { pageSize, pageToken } = getPaginationParams(requestUrl);
-  const { paginatedData, nextPageToken } = getPaginatedResponse(
+  const { pageSize, pageToken } = getDebArchivePaginationParams(requestUrl);
+  const { paginatedData, nextPageToken } = getDebArchivePaginatedResponse(
     locals,
     pageToken,
     pageSize,
@@ -104,8 +55,8 @@ const getLocalsResponse = (requestUrl: string) => {
 };
 
 const getPublicationTargetsResponse = (requestUrl: string) => {
-  const { pageSize, pageToken } = getPaginationParams(requestUrl);
-  const { paginatedData, nextPageToken } = getPaginatedResponse(
+  const { pageSize, pageToken } = getDebArchivePaginationParams(requestUrl);
+  const { paginatedData, nextPageToken } = getDebArchivePaginatedResponse(
     publicationTargets,
     pageToken,
     pageSize,
@@ -120,7 +71,7 @@ const getPublicationTargetsResponse = (requestUrl: string) => {
 const getPublicationsResponse = (requestUrl: string) => {
   const { searchParams } = new URL(requestUrl);
   const search = searchParams.get("search") ?? undefined;
-  const { pageSize, pageToken } = getPaginationParams(requestUrl);
+  const { pageSize, pageToken } = getDebArchivePaginationParams(requestUrl);
 
   const filteredPublications = search
     ? generateFilteredResponse(publications, search, [
@@ -130,7 +81,7 @@ const getPublicationsResponse = (requestUrl: string) => {
       ])
     : publications;
 
-  const { paginatedData, nextPageToken } = getPaginatedResponse(
+  const { paginatedData, nextPageToken } = getDebArchivePaginatedResponse(
     filteredPublications,
     pageToken,
     pageSize,
@@ -220,17 +171,6 @@ const getBatchPublicationTargetsResponse = async (
   return HttpResponse.json({ publicationTargets: matched });
 };
 
-const getBatchMirrorsResponse = async (
-  request: Request,
-): Promise<StrictResponse<BatchGetMirrorsResponse>> => {
-  const body = (await request.json()) as { names?: string[] };
-  const requestedNames = body.names ?? [];
-  const matched = mirrors.filter(({ name }) =>
-    name ? requestedNames.includes(name) : false,
-  );
-  return HttpResponse.json({ mirrors: matched });
-};
-
 const getBatchLocalsResponse = async (
   request: Request,
 ): Promise<StrictResponse<BatchGetLocalsResponse>> => {
@@ -250,72 +190,8 @@ export default [
     },
   ),
 
-  http.post(`${API_URL_DEB_ARCHIVE}mirrors:batchGet`, async ({ request }) => {
-    return getBatchMirrorsResponse(request);
-  }),
-
   http.post(`${API_URL_DEB_ARCHIVE}locals:batchGet`, async ({ request }) => {
     return getBatchLocalsResponse(request);
-  }),
-
-  http.get(`${API_URL_DEB_ARCHIVE}mirrors`, async ({ request }) => {
-    await delay();
-
-    const endpointStatus = getEndpointStatus();
-
-    if (
-      endpointStatus.status === "error" &&
-      endpointStatus.path === "mirrors"
-    ) {
-      return ENDPOINT_STATUS_API_ERROR;
-    }
-
-    return getMirrorsResponse(request.url);
-  }),
-
-  http.post(`${API_URL_DEB_ARCHIVE}mirrors`, async () => {
-    await delay();
-    return HttpResponse.json<CreateMirrorResponse>();
-  }),
-
-  http.get(`${API_URL_DEB_ARCHIVE}mirrors/:mirrorName`, async () => {
-    await delay();
-    return HttpResponse.json<GetMirrorResponse>(mirrors[0]);
-  }),
-
-  http.get(`${API_URL_DEB_ARCHIVE}mirrors/:mirrorName/packages`, async () => {
-    await delay();
-    return HttpResponse.json<SyncMirrorResponse>();
-  }),
-
-  http.patch(`${API_URL_DEB_ARCHIVE}mirrors/:mirrorName`, async () => {
-    await delay();
-    return HttpResponse.json<UpdateMirrorResponse>();
-  }),
-
-  http.delete(`${API_URL_DEB_ARCHIVE}mirrors/:mirrorName`, async () => {
-    await delay();
-    return HttpResponse.json<DeleteMirrorResponse>();
-  }),
-
-  http.post(`${API_URL_DEB_ARCHIVE}mirrors/:mirrorName\\:sync`, async () => {
-    await delay();
-
-    const endpointStatus = getEndpointStatus();
-
-    if (
-      endpointStatus.status === "error" &&
-      matchesPublicationsPath(endpointStatus.path)
-    ) {
-      return HttpResponse.json(
-        { message: "Failed to get packages" },
-        { status: 500 },
-      );
-    }
-
-    return HttpResponse.json<ListMirrorPackagesResponse>({
-      mirrorPackages: ["package-1", "package-2", "package-3"],
-    });
   }),
 
   http.get(`${API_URL_DEB_ARCHIVE}locals`, ({ request }) => {
